+++ apps/deploy-web/src/components/deployments/CreateCertificateButton/CreateCertificateButton.tsx
@@ -1,41 +1,46 @@
 import type { FC } from "react";
-import { useCallback } from "react";
+import { useCallback, useMemo } from "react";
 import type { ButtonProps } from "@akashnetwork/ui/components";
 import { Alert, Button, Spinner } from "@akashnetwork/ui/components";
 import { cn } from "@akashnetwork/ui/utils";
 
 import { useCertificate } from "@src/context/CertificateProvider";
 
-export const COMPONENTS = {
+export const DEPENDENCIES = {
   Alert,
-  Button
+  Button,
+  Spinner,
+  useCertificate
 };
 
 export interface Props extends Omit<ButtonProps, "onClick"> {
-  warningText?: string;
   afterCreate?: () => void;
   containerClassName?: string;
-  components?: typeof COMPONENTS;
+  dependencies?: typeof DEPENDENCIES;
 }
 
-export const CreateCertificateButton: FC<Props> = ({ warningText, afterCreate, containerClassName, components: c = COMPONENTS, ...buttonProps }) => {
-  const { isCreatingCert, createCertificate } = useCertificate();
+export const CreateCertificateButton: FC<Props> = ({ afterCreate, containerClassName, dependencies: d = DEPENDENCIES, ...buttonProps }) => {
+  const { isCreatingCert, createCertificate, isLocalCertExpired, localCert } = d.useCertificate();
 
   const _createCertificate = useCallback(async () => {
     await createCertificate();
     afterCreate?.();
   }, [createCertificate, afterCreate]);
+  const warningText = useMemo(() => {
+    if (isLocalCertExpired) return "Your certificate has expired. Please create a new one.";
+    if (!localCert) return "You need to create a certificate to view deployment details.";
+    return undefined;
+  }, [isLocalCertExpired, isCreatingCert, localCert]);
+  const buttonText = useMemo(() => (isLocalCertExpired ? "Regenerate Certificate" : "Create Certificate"), [isLocalCertExpired]);
 
   return (
     <div className={containerClassName}>
-      {warningText && (
-        <c.Alert variant="warning" className={cn({ "py-2 text-sm": buttonProps?.size === "sm" }, "truncate")}>
-          {warningText}
-        </c.Alert>
-      )}
-      <c.Button className={warningText ? "mt-4" : ""} {...buttonProps} disabled={buttonProps?.disabled || isCreatingCert} onClick={_createCertificate}>
-        {isCreatingCert ? <Spinner size="small" /> : "Create Certificate"}
-      </c.Button>
+      <d.Alert variant="warning" className={cn({ "py-2 text-sm": buttonProps?.size === "sm" }, "truncate")}>
+        {warningText}
+      </d.Alert>
+      <d.Button className={warningText ? "mt-4" : ""} {...buttonProps} disabled={buttonProps?.disabled || isCreatingCert} onClick={_createCertificate}>
+        {isCreatingCert ? <d.Spinner size="small" /> : buttonText}
+      </d.Button>
     </div>
   );
 };