+++ apps/deploy-web/src/components/new-deployment/NewDeploymentContainer.tsx
@@ -17,7 +17,7 @@ import { RouteStep } from "@src/types/route-steps.type";
 import { hardcodedTemplates } from "@src/utils/templates";
 import { UrlService } from "@src/utils/urlUtils";
 import Layout from "../layout/Layout";
-import { CreateLease } from "./CreateLease";
+import { CreateLease } from "./CreateLease/CreateLease";
 import { ManifestEdit } from "./ManifestEdit";
 import { CustomizedSteppers } from "./Stepper";
 import { TemplateList } from "./TemplateList";