+++ apps/deploy-web/src/components/wallet/ConnectManagedWalletButton.tsx
@@ -5,10 +5,11 @@ import type { ButtonProps } from "@akashnetwork/ui/components";
 import { Button, Spinner } from "@akashnetwork/ui/components";
 import { cn } from "@akashnetwork/ui/utils";
 import { Rocket } from "iconoir-react";
+import { useRouter } from "next/navigation";
 
 import { useWallet } from "@src/context/WalletProvider";
-import { useLoginRequiredEventHandler } from "@src/hooks/useLoginRequiredEventHandler";
-import { useFeatureFlags } from "@src/queries/featureFlags";
+import { useFlag } from "@src/hooks/useFlag";
+import { UrlService } from "@src/utils/urlUtils";
 
 interface Props extends ButtonProps {
   children?: ReactNode;
@@ -17,18 +18,16 @@ interface Props extends ButtonProps {
 
 export const ConnectManagedWalletButton: React.FunctionComponent<Props> = ({ className = "", ...rest }) => {
   const { connectManagedWallet, hasManagedWallet, isWalletLoading } = useWallet();
-  const { data: features } = useFeatureFlags();
-  const whenLoggedIn = useLoginRequiredEventHandler();
-  const startTrial: React.MouseEventHandler = useCallback(
-    event => {
-      if (features?.allowAnonymousUserTrial) {
-        connectManagedWallet();
-      } else {
-        whenLoggedIn("Sign In or Sign Up to start trial")(connectManagedWallet)(event);
-      }
-    },
-    [connectManagedWallet, features?.allowAnonymousUserTrial, whenLoggedIn]
-  );
+  const allowAnonymousUserTrial = useFlag("anonymous_free_trial");
+  const router = useRouter();
+
+  const startTrial: React.MouseEventHandler = useCallback(() => {
+    if (allowAnonymousUserTrial) {
+      connectManagedWallet();
+    } else {
+      router.push(UrlService.onboarding());
+    }
+  }, [connectManagedWallet, allowAnonymousUserTrial, router]);
 
   return (
     <Button