+++ apps/deploy-web/src/queries/useBlocksQuery.ts
@@ -20,10 +20,10 @@ export function useBlock(id: string, options = {}) {
 }
 
 export function useBlocks(limit: number, options?: Omit<UseQueryOptions<Block[], Error, any, QueryKey>, "queryKey" | "queryFn">) {
-  const { axios } = useServices();
+  const { publicConsoleApiHttpClient } = useServices();
   return useQuery<Block[], Error>({
     queryKey: QueryKeys.getBlocksKey(limit),
-    queryFn: () => axios.get(ApiUrlService.blocks(limit)).then(response => response.data),
+    queryFn: () => publicConsoleApiHttpClient.get(ApiUrlService.blocks(limit)).then(response => response.data),
     ...options
   });
 }