+++ apps/api/CHANGELOG.md
@@ -1,5 +1,49 @@
 
 
+## [2.108.1](https://github.com/akash-network/console/compare/console-api/v2.108.0...console-api/v2.108.1) (2025-07-28)
+
+
+### Code Refactoring
+
+* removes old implementation of feature flags based on env variables ([#1480](https://github.com/akash-network/console/issues/1480)) ([fa722aa](https://github.com/akash-network/console/commit/fa722aae1baf9e5dbd034381a2b641cfd7a5bf8d))
+
+## [2.108.0](https://github.com/akash-network/console/compare/console-api/v2.107.2...console-api/v2.108.0) (2025-07-27)
+
+
+### Features
+
+* adds safe node packages installation ([#1726](https://github.com/akash-network/console/issues/1726)) ([37acfee](https://github.com/akash-network/console/commit/37acfee5c1d053cec2316560ad220992d70b7cbf)), closes [#1549](https://github.com/akash-network/console/issues/1549)
+* **billing:** new onboarding flow ([#1711](https://github.com/akash-network/console/issues/1711)) ([e05506c](https://github.com/akash-network/console/commit/e05506c89a762e8fa9111649357aa80c06e26fbe))
+* **deployment:** close trial deployments after 24h ([#1725](https://github.com/akash-network/console/issues/1725)) ([05017ac](https://github.com/akash-network/console/commit/05017acb217be1efdfbe545f868f9c8f816e718d))
+
+
+### Bug Fixes
+
+* **auth:** fetching api keys ([#1743](https://github.com/akash-network/console/issues/1743)) ([cda80ce](https://github.com/akash-network/console/commit/cda80cefacff2d677fdd4e334b9d1997b6b9bd95))
+* set a better date to work with ([#1729](https://github.com/akash-network/console/issues/1729)) ([218cdc4](https://github.com/akash-network/console/commit/218cdc4578d847338dd41945beb2fd29c19b0466))
+
+## [2.107.2](https://github.com/akash-network/console/compare/console-api/v2.107.1...console-api/v2.107.2) (2025-07-22)
+
+
+### Bug Fixes
+
+* fix wallet creation error on stripe webhook ([#1684](https://github.com/akash-network/console/issues/1684)) ([595275e](https://github.com/akash-network/console/commit/595275e2cf8b749d1c404759d397d4d1819eaefd)), closes [#1662](https://github.com/akash-network/console/issues/1662)
+* make userAgent 500 max chars long ([#1677](https://github.com/akash-network/console/issues/1677)) ([b929e99](https://github.com/akash-network/console/commit/b929e99af69d88ba2c1c8cd39c1b19fe1f1334ad)), closes [#1520](https://github.com/akash-network/console/issues/1520)
+
+## [2.107.1](https://github.com/akash-network/console/compare/console-api/v2.107.0...console-api/v2.107.1) (2025-07-22)
+
+
+### Code Refactoring
+
+* creates separate axios instance per intent ([#1702](https://github.com/akash-network/console/issues/1702)) ([b15370f](https://github.com/akash-network/console/commit/b15370fccc676982705c41c7b659752a467aef0d)), closes [#1668](https://github.com/akash-network/console/issues/1668)
+
+## [2.107.0](https://github.com/akash-network/console/compare/console-api/v2.106.1...console-api/v2.107.0) (2025-07-21)
+
+
+### Features
+
+* **provider:** added provider-earnings api ([#1646](https://github.com/akash-network/console/issues/1646)) ([3376863](https://github.com/akash-network/console/commit/3376863c1f817c73438a917bec6ada5f0031a2b7))
+
 ## [2.106.1](https://github.com/akash-network/console/compare/console-api/v2.106.0...console-api/v2.106.1) (2025-07-16)